Refactored admin dashboard code to split styling/presentation/behavioural concerns
