+++ App/components/UnassignedEventList/styles.js
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+export default styles = StyleSheet.create({
+  container: {
+    flex: 1
+  }
+});
