+++ App/views/Home/AdminDashboard/index.js
@@ -1,11 +1,6 @@
 import React, { Component } from 'react';
-import {
-  FlatList,
-  StyleSheet,
-  Text,
-  View
-} from 'react-native';
-import UnassignedEventListContainer from '../../../components/UnassignedEventList/';
+import { View } from 'react-native';
+import UnassignedEventList from '../../../components/UnassignedEventList';
 
 export default class AdminDashboard extends Component {
   static navigationOptions = {
@@ -13,22 +8,8 @@ export default class AdminDashboard extends Component {
   };
 
   render() {
-    return (
-        <View style={styles.container}>
-            <UnassignedEventListContainer />
-        </View>
-    );
+    return <View style={styles.container}>
+      <UnassignedEventList />
+    </View>;
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F5FCFF',
-  },
-  title: {
-    fontSize: 24,
-    textAlign: 'center',
-    margin: 15,
-  },
-});
